backend/requirements.txt
fastapi>=0.128.0
psycopg2-binary>=2.9.11
pydantic>=2.12.5
pyjwt>=2.10.1
cryptography>=41.0.0
python-dotenv>=1.2.1
python-telegram-bot>=22.5
sqlalchemy>=2.0.46
uvicorn[standard]>=0.40.0
httpx>=0.27.0
openai>=1.59.0
